test: cover placement denylist helpers
